requirements.txt
moviepy==2.1.2
ipython==9.0.2
opencv-python==4.11.0.86
pillow==10.4.0
google-generativeai==0.8.4
streamlit==1.43.2